backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
pymysql==1.1.0
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0